[Cleanup] Move Setup Guide images
- to be nested within their own component views that use them, instead of having to grab them upwards from a shared assets folder
