+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/setup_guide/assets/getting_started.png
