+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/views/setup_guide/assets/getting_started.png
